+++ aws-sdk-core/lib/aws-sdk-core/iot.rb
@@ -1,0 +1,4 @@
+Aws.add_service(:IoT, {
+  api: "#{Aws::API_DIR}/iot/2015-05-28/api-2.json",
+  docs: "#{Aws::API_DIR}/iot/2015-05-28/docs-2.json",
+})